Change BigDecimal constructors to use double. This should make the compile with JDK 1.5 work with Java 1.4 and before. Fixes TORQUE-97
git-svn-id: http://svn.apache.org/repos/asf/db/torque/runtime/trunk@552333 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/torque/oid/IDBroker.java
@@ -744,7 +744,7 @@ public class IDBroker implements Runnable, IdGenerator
         // If prefetch is turned off we simply return 1
         if (!configuration.getBoolean(DB_IDBROKER_PREFETCH, true))
         {
-            quantity = new BigDecimal(1);
+            quantity = new BigDecimal((double)1);
         }
         // Initialize quantity, if necessary.
         else if (quantityStore.containsKey(tableName))
@@ -772,7 +772,7 @@ public class IDBroker implements Runnable, IdGenerator
             }
             catch (Exception e)
             {
-                quantity = new BigDecimal(10);
+                quantity = new BigDecimal((double)10);
             }
             finally
             {
